feat: Refactor chat and message models, add read receipts
This commit refactors the chat and message models:
- Updates chat model to use `timestamp` field for message ordering.
- Introduces many-to-many relationship for message read receipts.
- Updates message service to update chat timestamp on new messages.
- Allows message deletion by setting text to "deleted message".
- Updates message deletion notification to support multiple opponents.
- Adds relation for seen messages to the User model.

+++ src/app/modules/message/Message.service.ts
@@ -15,15 +15,36 @@ export const MessageServices = {
    * Create new message
    */
   async createMessage(payload: TCreateMessageArgs) {
-    return prisma.message.create({
-      data: payload,
-      include: {
-        chat: {
-          select: {
-            user_ids: true,
+    return prisma.$transaction(async tx => {
+      //? update chat timestamp
+      tx.chat.update({
+        where: { id: payload.chat_id },
+        data: { timestamp: new Date() },
+      });
+
+      //? create message
+      return tx.message.create({
+        data: {
+          ...payload,
+          seen_by: {
+            connect: {
+              id: payload.user_id,
+            },
           },
         },
-      },
+        include: {
+          chat: {
+            select: {
+              user_ids: true,
+            },
+          },
+          seen_by: {
+            select: {
+              avatar: true,
+            },
+          },
+        },
+      });
     });
   },
 
@@ -33,7 +54,7 @@ export const MessageServices = {
   async deleteMessage({ message_id, user_id }: TDeleteMessageArgs) {
     const message = await prisma.message.findUnique({
       where: { id: message_id },
-      select: { user_id: true, media_urls: true },
+      select: { user_id: true, media_urls: true, isDeleted: true },
     });
 
     //? ensure that user has permission to delete message
@@ -44,15 +65,22 @@ export const MessageServices = {
       );
     }
 
+    if (message.isDeleted) {
+      throw new ServerError(StatusCodes.BAD_REQUEST, 'Message already deleted');
+    }
+
     await deleteFiles(message.media_urls);
 
-    return prisma.message.delete({
+    return prisma.message.update({
       where: { id: message_id },
+      data: {
+        media_urls: [],
+        text: '𝒹𝑒𝓁𝑒𝓉𝑒𝒹 𝓂𝑒𝓈𝓈𝒶𝑔𝑒',
+        isDeleted: true,
+      },
       select: {
         chat: {
-          select: {
-            user_ids: true,
-          },
+          select: { id: true, user_ids: true },
         },
       },
     });
